# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "House of Evolve API",
    "version": "1.0.0",
    "description": "Multi-brand catalog API for IRA and Jerseymise, plus combined catalog for HOE."
  },
  "servers": [
    { "url": "https://hoe-be.onrender.com", "description": "Production server" },
    { "url": "http://localhost:{port}", "variables": { "port": { "default": "4000" } }, "description": "Local development server" }
  ],
  "paths": {
...
    "/api/brands/{brandSlug}/categories": {
      "get": {
        "summary": "List categories for a brand",
        "parameters": [
          { "name": "brandSlug", "in": "path", "required": true, "schema": { "type": "string", "example": "ira" } }
        ],
        "responses": { "200": { "description": "List of categories" }, "404": { "description": "Brand not found" } }
      }
    },
    "/api/brands/{brandSlug}/categories/{categorySlug}/subcategories": {
      "get": {
        "summary": "List subcategories for a category in a brand",
        "parameters": [
          { "name": "brandSlug", "in": "path", "required": true, "schema": { "type": "string", "example": "ira" } },
          { "name": "categorySlug", "in": "path", "required": true, "schema": { "type": "string", "example": "necklaces" } }
        ],
        "responses": { "200": { "description": "List of subcategories" }, "404": { "description": "Not found" } }
      }
    },
    "/api/brands/{brandSlug}/products": {
      "get": {
        "summary": "List products for a brand",
        "description": "Retrieve products for a specific brand with populated category and subcategory data. Supports filtering, searching, and pagination.",
        "tags": ["Brands", "Products"],
        "parameters": [
          { "name": "brandSlug", "in": "path", "required": true, "schema": { "type": "string", "example": "jerseymise" } },
          { "name": "category", "in": "query", "schema": { "type": "string", "description": "Filter by category slug" } },
          { "name": "subcategory", "in": "query", "schema": { "type": "string", "description": "Filter by subcategory slug" } },
          { "name": "q", "in": "query", "schema": { "type": "string", "description": "Search query for title, description, or tags" } },
          { "name": "min", "in": "query", "schema": { "type": "number", "description": "Minimum price filter" } },
          { "name": "max", "in": "query", "schema": { "type": "number", "description": "Maximum price filter" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1, "description": "Page number for pagination" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "description": "Number of items per page" } }
        ],
        "responses": { 
          "200": { 
            "description": "Paginated products with populated data", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Product" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }, 
          "404": { "description": "Brand not found" } 
        }
      }
    },
...
    "/api/brands/{brandSlug}/categories/{categorySlug}/products": {
      "get": {
        "summary": "List products for a brand category",
        "description": "Retrieve products for a specific brand and category with populated subcategory data.",
        "tags": ["Brands", "Products"],
        "parameters": [
          { "name": "brandSlug", "in": "path", "required": true, "schema": { "type": "string", "example": "jerseymise" } },
          { "name": "categorySlug", "in": "path", "required": true, "schema": { "type": "string", "example": "shirts" } }
        ],
        "responses": { 
          "200": { 
            "description": "Products for the specified brand and category with populated data", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Product" }
                    }
                  }
                }
              }
            }
          }, 
          "404": { "description": "Brand or category not found" } 
        }
      }
    },
    "/api/brands/{brandSlug}/categories/{categorySlug}/subcategories/{subcategorySlug}/products": {
      "get": {
        "summary": "List products for a brand subcategory",
        "description": "Retrieve products for a specific brand, category, and subcategory with populated data.",
        "tags": ["Brands", "Products"],
        "parameters": [
          { "name": "brandSlug", "in": "path", "required": true, "schema": { "type": "string", "example": "jerseymise" } },
          { "name": "categorySlug", "in": "path", "required": true, "schema": { "type": "string", "example": "shirts" } },
          { "name": "subcategorySlug", "in": "path", "required": true, "schema": { "type": "string", "example": "polo" } }
        ],
        "responses": { 
          "200": { 
            "description": "Products for the specified brand, category, and subcategory with populated data", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Product" }
                    }
                  }
                }
              }
            }
          }, 
          "404": { "description": "Brand, category, or subcategory not found" } 
        }
      }
    },
...
    "/api/admin/brands": {
      "post": {
        "summary": "Create a brand (admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Brand" } } } },
          "409": { "description": "Slug already exists" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/brands/{id}": {
      "patch": {
        "summary": "Update a brand (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Brand" } } } },
          "404": { "description": "Brand not found" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Delete a brand (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Brand not found" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/brands/{brandId}/categories": {
      "post": {
        "summary": "Create category for a brand (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "brandId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "image": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "404": { "description": "Brand not found" },
          "409": { "description": "Slug already exists for this brand" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
    "/api/admin/brands/{brandId}/categories/{categoryId}/subcategories": {
      "post": {
        "summary": "Create subcategory for a category (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "brandId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "categoryId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "image": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subcategory" } } } },
          "404": { "description": "Brand or Category not found" },
          "409": { "description": "Slug already exists for this category" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "brandId": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "image": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        }
      },
      "Brand": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "active": { "type": "boolean" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        }
      },
...
      "Pagination": { "type": "object", "properties": { "page": { "type": "integer" }, "pages": { "type": "integer" }, "total": { "type": "integer" } } },
...
    },
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
  }
}